AI_vs_Real_Deployment/requirements.txt
streamlit==1.29.0
tensorflow-cpu==2.14.0
numpy==1.25.2
pillow==10.0.0